// frontend/src/types/elements.type.ts
export type ElementsType = {
    typeSelect: HTMLSelectElement | null;
    categorySelect: HTMLSelectElement | null;
    amountInput: HTMLInputElement | null;
    dateInput: HTMLInputElement | null;
    commentInput: HTMLInputElement | null;
    saveButton: HTMLButtonElement | null;
    cancelButton: HTMLButtonElement | null;
    errorContainer: HTMLElement | null;
}